requirements.txt
tensorflow
opencv-python-headless
pillow
numpy
fastapi
uvicorn
python-multipart
